backend/requirements.txt
pytest
alembic
annotated-types
anyio
bcrypt
cffi
click
colorama
cryptography
ecdsa
fastapi
greenlet
h11
idna
Mako
MarkupSafe
passlib
psycopg2
pyasn1
pycparser
pydantic
pydantic_core
python-dotenv
python-jose
python-multipart
redis
rsa
six
sniffio
SQLAlchemy
starlette
typing_extensions
uvicorn
httpx